+++ src/main/java/com/traders/traders/module/trading/strategy/domain/Strategy.java
@@ -1,4 +1,4 @@
-package com.traders.traders.module.trading.strategy;
+package com.traders.traders.module.trading.strategy.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Embedded;
